Remove annoying white dots. Closes #3
This turned out to be something called 'texture bleeding'. It's caused
when the S/T coordinates end up bordering the white in the texture JPG.
The way I chose to fix it just nudges the texture S/T coordinates
slightly inwards on the texture, so that you don't end up sampling
colors just off the color.